Update and Rename
I updated some of my SignUp.php code and renamed the newUser code to a different name for readability purposes
